# backend/openapi.yaml
components:
...
  schemas:
...
    BinaryResponse:
      properties:
        message:
          example: Failed load data
          type: string
        result:
          example: false
          type: boolean
      type: object
...
    FileMasking:
      properties:
        description:
          type: string
        filename:
          type: string
        id:
          type: integer
        is_test:
          type: boolean
        is_valid:
          type: boolean
        masking_uuid:
          format: uuid
          type: string
        user_id:
          type: integer
      type: object
...
    GenerateMaskingPlan:
      properties:
        is_test:
          type: boolean
        locations:
          items:
            $ref: '#/components/schemas/LocationGenerateMap'
          type: array
        questions:
          items:
            $ref: '#/components/schemas/Question'
          type: array
        type_works:
          items:
            $ref: '#/components/schemas/TypeWorkGenerateMap'
          nullable: true
          type: array
      type: object
    ListFileMasking:
      properties:
        files:
          items:
            $ref: '#/components/schemas/FileMasking'
          type: array
        pagination:
          $ref: '#/components/schemas/PaginationResponse'
      type: object
...
    LocationGenerateMap:
      properties:
        id:
          type: integer
        id_type:
          nullable: true
          type: integer
      type: object
...
    PaginationResponse:
      properties:
        limit:
          example: 10
          type: integer
        page:
          example: 1
          type: integer
        total_items:
          example: 10
          type: integer
        total_pages:
          example: 1
          type: integer
      type: object
...
    Question:
      properties:
        answer_id:
          type: integer
        id:
          type: integer
      type: object
...
    TypeWorkGenerateMap:
      properties:
        id:
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
    roleAuth:
      in: header
      name: Role
      type: apiKey
    userAuth:
      in: header
      name: User
      type: apiKey
info:
  title: "\u041F\u0440\u0438\u043B\u043E\u0436\u0435\u043D\u0438\u0435 \u0434\u043B\
    \u044F \u0433\u0435\u043D\u0435\u0440\u0430\u0446\u0438\u0438 \u043A\u0430\u0440\
    \u0442 \u043C\u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\u0438\u044F"
  version: v0.1
openapi: 3.0.3
paths:
...
  /api/files/get-pdf/:
    get:
      parameters:
      - description: "\u041F\u0430\u0440\u0430\u043C\u0435\u0442\u0440\u044B"
        in: query
        name: masking_uuid
        type: str
      responses:
        '200':
          content:
            app/pdf:
              schema:
                type: file
          description: "\u0424\u0430\u0439\u043B"
        '400':
          content:
            app/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u0424\u0430\u0439\u043B \u043D\u0435 \u0441\u0443\u0449\u0435\
            \u0441\u0442\u0432\u0443\u0435\u0442"
      summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0444\u0430\u0439\
        \u043B"
      tags:
      - map_files
  /api/files/get-html/:
    get:
      parameters:
      - description: "\u041F\u0430\u0440\u0430\u043C\u0435\u0442\u0440\u044B"
        in: query
        name: masking_uuid
        type: str
      responses:
        '200':
          content:
            text/html:
              schema:
                type: string
          description: "\u0424\u0430\u0439\u043B"
        '400':
          content:
            app/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u0424\u0430\u0439\u043B \u043D\u0435 \u0441\u0443\u0449\u0435\
            \u0441\u0442\u0432\u0443\u0435\u0442"
      summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0438\u043D\u0444\
        \u043E\u0440\u043C\u0430\u0446\u0438\u044E \u043F\u043E \u043C\u0430\u043A\
        \u0441\u0438\u0440\u043E\u0432\u0430\u043D\u0438\u044E"
      tags:
      - map_files
  /api/files/get-files/:
    get:
      description: "\u0424\u0438\u043B\u044C\u0442\u0440\u0430\u0446\u0438\u044F \u0441\
        \u043F\u0438\u0441\u043A\u0430 \u0444\u0430\u0439\u043B\u043E\u0432 \u043A\
        \u0430\u0440\u0442 \u043C\u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\
        \u0438\u044F"
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 1
          example: 1
          minimum: 0
          type: integer
      - explode: true
        in: query
        name: protection_ids[]
        required: false
        schema:
          items:
            type: integer
          nullable: true
          type: array
        style: form
      - explode: true
        in: query
        name: type_location_ids[]
        required: false
        schema:
          items:
            type: integer
          nullable: true
          type: array
        style: form
      - in: query
        name: limit
        required: false
        schema:
          default: 1000
          example: 10
          type: integer
      - explode: true
        in: query
        name: type_work_ids[]
        required: false
        schema:
          items:
            type: integer
          nullable: true
          type: array
        style: form
      responses:
        '200':
          content:
            app/json:
              schema:
                $ref: '#/components/schemas/ListFileMasking'
          description: "\u0421\u043F\u0438\u0441\u043E\u043A \u043A\u0430\u0440\u0442"
        '400':
          content:
            app/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u041E\u0448\u0438\u0431\u043A\u0430 \u043F\u0440\u0438 \u043F\
            \u043E\u043B\u0443\u0447\u0435\u043D\u0438\u0438"
      summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u043E\u0442\u0444\
        \u0438\u043B\u044C\u0442\u0440\u043E\u0432\u0430\u043D\u043D\u044B\u0439 \u0441\
        \u043F\u0438\u0441\u043E\u043A \u043A\u0430\u0440\u0442"
      tags:
      - map_files
  /api/files/generate-masking/:
    post:
      description: "\u0413\u0435\u043D\u0435\u0440\u0430\u0446\u0438\u044F \u043A\u0430\
        \u0440\u0442\u044B \u043C\u0430\u0441\u043A\u0438\u0440\u043E\u0432\u0430\u043D\
        \u0438\u044F"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateMaskingPlan'
        description: "\u041F\u0430\u0440\u0430\u043C\u0435\u0442\u0440\u044B"
      responses:
        '200':
          content:
            app/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u0420\u0435\u0437\u0443\u043B\u044C\u0442\u0430\u0442 \u043F\
            \u0440\u043E\u0432\u0435\u0440\u043A\u0438"
        '400':
          content:
            app/json:
              schema:
                $ref: '#/components/schemas/BinaryResponse'
          description: "\u041A\u0430\u0440\u0442\u0430 \u043D\u0435 \u043D\u0443\u0436\
            \u043D\u0430"
      summary: "\u0421\u0433\u0435\u043D\u0435\u0440\u0438\u0440\u043E\u0432\u0430\
        \u0442\u044C \u043A\u0430\u0440\u0442\u0443 \u043C\u0430\u0441\u043A\u0438\
        \u0440\u043E\u0432\u0430\u043D\u0438\u044F"
      tags:
      - map_files
...
servers:
- description: "\u0422\u0435\u0441\u0442\u043E\u0432\u044B\u0439 \u0441\u0435\u0440\
    \u0432\u0435\u0440"
  url: http://localhost:5001/